requirements.txt
python==3.9.13
python-telegram-bot==20.0a4
mysql-connector-python==8.0.31